Move Transaction code to separate files
