+++ src/realm/sync/noinst/client_reset_operation.cpp
@@ -16,7 +16,7 @@
  *
  **************************************************************************/
 
-#include <realm/db.hpp>
+#include <realm/transaction.hpp>
 #include <realm/sync/history.hpp>
 #include <realm/sync/noinst/client_history_impl.hpp>
 #include <realm/sync/noinst/client_reset.hpp>
