+++ src/realm/sync/noinst/sync_metadata_schema.cpp
@@ -19,7 +19,7 @@
 #include "realm/sync/noinst/sync_metadata_schema.hpp"
 
 #include "realm/data_type.hpp"
-#include "realm/db.hpp"
+#include "realm/transaction.hpp"
 #include "realm/util/flat_map.hpp"
 #include "realm/util/optional.hpp"
 #include <stdexcept>
