+++ test/test_links.cpp
@@ -23,7 +23,6 @@
 #include <realm.hpp>
 #include <realm/util/file.hpp>
 #include <realm/array_key.hpp>
-#include <realm/history.hpp>
 
 #include "test.hpp"
