+++ test/test_query.cpp
@@ -29,7 +29,6 @@ using namespace std::chrono;
 #include <realm.hpp>
 #include <realm/column_integer.hpp>
 #include <realm/array_bool.hpp>
-#include <realm/history.hpp>
 #include <realm/query_expression.hpp>
 #include <realm/index_string.hpp>
 #include <realm/query_expression.hpp>
